feat(GetArticle): adds new protocol
* Contains methods related to requesting a single article
* Moves getBy method form ListArticle to GetArticle
* Splits AppArticlesRestAdapterTests
  * Now there is a dedicated test for each protocol implemented by AppArticleRestAdapter
